Added JUnit4 test support, added a few basic tests for the Logger class.
See comment in TeamCode package's build.gradle file - had to add something so that it found Junit4. Now we can run JUnit tests on the Desktop for our robotics code! Will need to add mock classes for FTC-specific dependencies, but now we have a very clean way to do all of this. Yay!