Partial Mock Hardware Setup
Defined most of the hardware we'll likely be using: motors, servos, a few sensors, analog input and output, etc.
Created new interface HardwareLookupInterface - this will take the place of HardwreMap, which is a very messy class, so very hard to mock.
Created Mock hardware and mock Hardware lookup - under test.
None of this code runs (yet) but it compiles.