Mock testing of Lights!
ITask_Lights is the new-format task (ITask stands for Iterative Task). I has been tested in a nontrivial manner by ITask_LightsTest. MockAnalogOutputController now keeps track of the state of each analog output port. Tests verify that the voltage and mode as as expected based on the current (simulated) time.