Red-alliance Auton code - mostly working
Aparna and Keya coded this up.
This code has the logic to work for both left and right stones of the Red alliance. It currently doesn't read the color sensor, but the logic is present to do the right thing when the sensor returns RED/BLUE/UNKNOWN.

This code was first tested on the LabBot and then on the Kiwibot (with a temporary servo connected - not connected to the wand as the wand servo could not reach the Rev hub.