Mira and Joseph - tweaks to encoderDrive timeout and color saturation
This is the "final" version for the 2nd competition - December 3rd.
encoderDrive had a timeout of 5.0 for the long stretch - that timeout was hitting, so bumped the timeout from 5.0 to 10.0. Also, the blue jewel was sometimes hitting a low saturation value - of 0.28, so bumped the lower limit to 0.15.

This version seems to work reliably for both colored jewels on both alliances and both slabs.