Added some temporary functionality to RawSensorModule
RawSensorModule reports color sensor values that are presently Perlin noise
that is correlated with the location of the robot on the field. This is a stopgap
until the querying of the field elements is complete, and also the use of Perlin noise
is a trial of its use to add noise to all sensors.